Add more tests specific to signing
